Split out grids into separate files
