+++ osu.Game.Rulesets.Osu.Tests/TestSceneOsuBeatSnapGrid.cs
@@ -7,16 +7,13 @@ using osu.Framework.Allocation;
 using osu.Framework.Graphics;
 using osu.Framework.Graphics.Containers;
 using osu.Framework.Graphics.Shapes;
-using osu.Framework.Graphics.UserInterface;
 using osu.Framework.Input.Events;
 using osu.Framework.MathUtils;
-using osu.Game.Beatmaps;
 using osu.Game.Beatmaps.ControlPoints;
-using osu.Game.Rulesets.Objects;
 using osu.Game.Rulesets.Osu.Beatmaps;
+using osu.Game.Rulesets.Osu.Edit;
 using osu.Game.Rulesets.Osu.Objects;
 using osu.Game.Screens.Edit;
-using osu.Game.Screens.Edit.Compose.Components;
 using osu.Game.Tests.Visual;
 using osuTK;
 using osuTK.Graphics;
@@ -207,81 +204,6 @@ namespace osu.Game.Rulesets.Osu.Tests
             public TestOsuBeatSnapGrid(OsuHitObject hitObject)
                 : base(hitObject)
             {
-            }
-        }
-
-        private abstract class CircularBeatSnapGrid : BeatSnapGrid
-        {
-            protected CircularBeatSnapGrid(HitObject hitObject, Vector2 centrePosition)
-                : base(hitObject, centrePosition)
-            {
-            }
-
-            protected override void CreateContent(Vector2 centrePosition)
-            {
-                float maxDistance = Math.Max(
-                    Vector2.Distance(centrePosition, Vector2.Zero),
-                    Math.Max(
-                        Vector2.Distance(centrePosition, new Vector2(DrawWidth, 0)),
-                        Math.Max(
-                            Vector2.Distance(centrePosition, new Vector2(0, DrawHeight)),
-                            Vector2.Distance(centrePosition, DrawSize))));
-
-                int requiredCircles = (int)(maxDistance / DistanceSpacing);
-
-                for (int i = 0; i < requiredCircles; i++)
-                {
-                    float radius = (i + 1) * DistanceSpacing * 2;
-
-                    AddInternal(new CircularProgress
-                    {
-                        Origin = Anchor.Centre,
-                        Position = centrePosition,
-                        Current = { Value = 1 },
-                        Size = new Vector2(radius),
-                        InnerRadius = 4 * 1f / radius,
-                        Colour = GetColourForBeatIndex(i)
-                    });
-                }
-            }
-
-            public override Vector2 GetSnapPosition(Vector2 position)
-            {
-                Vector2 direction = position - CentrePosition;
-                float distance = direction.Length;
-
-                float radius = DistanceSpacing;
-                int radialCount = Math.Max(1, (int)Math.Round(distance / radius));
-
-                if (radialCount <= 0)
-                    return position;
-
-                Vector2 normalisedDirection = direction * new Vector2(1f / distance);
-
-                return CentrePosition + normalisedDirection * radialCount * radius;
-            }
-        }
-
-        private class OsuBeatSnapGrid : CircularBeatSnapGrid
-        {
-            /// <summary>
-            /// Scoring distance with a speed-adjusted beat length of 1 second.
-            /// </summary>
-            private const float base_scoring_distance = 100;
-
-            public OsuBeatSnapGrid(OsuHitObject hitObject)
-                : base(hitObject, hitObject.StackedEndPosition)
-            {
-            }
-
-            protected override float GetVelocity(double time, ControlPointInfo controlPointInfo, BeatmapDifficulty difficulty)
-            {
-                TimingControlPoint timingPoint = controlPointInfo.TimingPointAt(time);
-                DifficultyControlPoint difficultyPoint = controlPointInfo.DifficultyPointAt(time);
-
-                double scoringDistance = base_scoring_distance * difficulty.SliderMultiplier * difficultyPoint.SpeedMultiplier;
-
-                return (float)(scoringDistance / timingPoint.BeatLength);
             }
         }
     }
